Support asset pipeline for loading CSS files
When running in Rails development mode using the asset pipeline and on a
single threaded server, CSS files were not getting properly loaded from
the rendered HTML. Further inspection determined that the `href`
attributes were from `link` tags were getting stripped by the
`premailer` gem (see `Premailer#convert_inline_links`,
https://github.com/premailer/premailer/blob/master/lib/premailer/premailer.rb#L412).
This is not an issue when in production as absolute URIs are given for
the assets by default, causing `Premailer` to load from the uri rather
than try to open the relative path locally.

The result is that an empty `link_uri` is used in
`Premailer#load_css_from_html!` when `link` tags are searched and
iterated over, causing load failures in development. The first work
around that needs to be added to a Rails project is to update your
environment file so that absolute URIs are generated by the asset
pipeline, allowing `Premailer` to pass on these hrefs when converting
inline links. I recommend that you update both your development and
production configuration as below:

```
config.action_mailer.asset_host = "http://your-domain.com"
```

Next, this patch overrides the `Premailer#load_css_from_html!` so that,
in development, we do the following:

1. Clean up the `link`'s href to remove the protocol, domain, asset
   prefix, and fingerprint.
2. Use Sprockets' `find_asset` to find the given asset.
3. Load the asset directly as a string (Sprockets has already any/all of
   the necessary preprocessing).

That that fix in place, I was able to get ActionMailer previews set up
with Premailer inlining CSS. The fix is a monkey-patched approach, which
is not ideal, and can use some refactoring. Unfortunately, `Premailer`
does not provide any seams where a cleaner approach could have been
taken. And, of course, knowledge of Rails' needs should be kept out of
`Premailer`.

I think that a cleaner upstream workaround could be to add an ability to
ignore certain URIs in `Premailer#convert_inline_links`.